+++ src/main/java/com/supermartijn642/movingelevators/elevator/ElevatorGroupCapability.java
@@ -96,7 +96,7 @@ public class ElevatorGroupCapability {
     }
 
     public void updateGroup(ElevatorGroup group){
-        if(!this.level.isClientSide && group != null)
+        if(!this.level.isClientSide() && group != null)
             MovingElevators.CHANNEL.sendToDimension(this.level.dimension(), new PacketAddElevatorGroup(this.writeGroup(group)));
     }
 
@@ -104,7 +104,7 @@ public class ElevatorGroupCapability {
         ElevatorGroup group = this.groups.remove(pos);
         if(group != null){
             this.groupsPerChunk.remove(pos.chunkPos(), group);
-            if(!this.level.isClientSide)
+            if(!this.level.isClientSide())
                 MovingElevators.CHANNEL.sendToDimension(this.level.dimension(), new PacketRemoveElevatorGroup(group));
         }
     }
@@ -113,7 +113,7 @@ public class ElevatorGroupCapability {
      * This should only be called client-side from the {@link PacketRemoveElevatorGroup}
      */
     public void removeGroup(int x, int z, Direction facing){
-        if(this.level.isClientSide)
+        if(this.level.isClientSide())
             this.removeGroup(new ElevatorGroupPosition(x, z, facing));
     }